agent_server/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Data Validation
pydantic==2.10.4
pydantic-settings>=2.0.0

# Environment
python-dotenv>=1.0.0

# HTTP Client (for local LLM)
httpx>=0.27.0

# LLM Providers
anthropic>=0.40.0
openai>=1.50.0
google-genai>=1.0.0

# Development/Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
